+++ authority-portal-frontend/src/app/pages/participant-own-connector-detail-page/participant-own-connector-detail-page/participant-own-connector-detail-page.component.ts
@@ -88,7 +88,7 @@ export class ParticipantOwnConnectorDetailPageComponent
         tap((state) => (this.state = state)),
         switchMap(() => {
           return this.globalStateUtils.userRoles$.pipe(
-            map((roles) => roles.has('PARTICIPANT_CURATOR')),
+            map((roles) => roles.has('KEY_USER')),
             distinctUntilChanged(),
           );
         }),